// src/openapi/models/ImageObject.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ImageObject = {
  /**
   * The source URL of the image.
 * 
   */
  url: string;
  /**
   * The image height in pixels.
 * 
   */
  height: number | null;
  /**
   * The image width in pixels.
 * 
   */
  width: number | null;
};